Fix error messages from reqparse
Before the request parser from Flask-RESTful was brought into
Flask-RESTPlus a commit was made to format the help messages of a parser
argument when creating the error string. This doesn't really work that
well with Flask-RESTPlus since those help messages are also used in the
swagger documentation. If someone wanted to retain this same
functionality another argument property would likely be needed. This
reverts things to how they worked previous to the merger.